+++ solution_db_service/src/main/java/cn/yangtengfei/repository/question/SolutionRepository.java
@@ -14,7 +14,7 @@ public interface SolutionRepository extends MongoRepository<Solution,String> {
 
     Page<Solution> findByQuestionIdOrderByUpdateTimeDesc(String questionId, Pageable pageable);
 
-    long countByQuestionId(String questionId);
+    Integer countByQuestionId(String questionId);
 
     List<Solution> findByQuestionIdAndDeleteFlg(String questionId,int deleteFlg);
